[OWL-3949] Sign-up auto reminder job classcastexception

+++ signup/impl/src/java/org/sakaiproject/signup/dao/SignupMeetingDao.java
@@ -229,7 +229,7 @@ public interface SignupMeetingDao extends GeneralGenericDao {
 	 * 			search ending date
 	 * @return the total record counts
 	 */
-	int getAutoReminderTotalEventCounts(Date startDate, Date endDate);
+	long getAutoReminderTotalEventCounts(Date startDate, Date endDate);
 	
 	/**
 	 * Get all the Categories from a site
